RG.test {chisqalt}	R Documentation

RG.test

Description

This function does a chi square goodness of fit test with binning chosen based on an alternative

Usage

RG.test(
  dta,
  LR = c(-Inf, Inf),
  pnull,
  param = NULL,
  palt,
  sample.sized.fixed = TRUE,
  nkappa = 5
)

Arguments

dta
(data to test)

pnull
cdf under the null

palt
cdf under the alternative

LR=c(-Inf,
Inf) support of distribution under the null

param=NULL
parameters of cdf under null, NULL if parameters are fixed

sample.size.fixed=TRUE
was sample sized fixed or from a Poisson distribution?

nkappa=5
(number of bin types to include)

Examples

 chisq.alt.test(dta=x, pnull=function(x, param) pnorm(x), palt=function(x) pt(x, 5))
 chisq.alt.test(dta=x, pnull=function(x, param) pnorm(x, param[1], param[2]), param=c(0, 1), palt=function(x) pt(x, 5))

[Package chisqalt version 0.0.0.900 Index]