power.multinomial {chisqalt}	R Documentation

power.multinomial

Description

This function finds power of the chisquare test using given binning and generating data using the multinomial distribution

Usage

power.multinomial(case, k, kappa, method = "Pearson", bins = NA)

Arguments

case

k

kappa

bins=NA

Examples

 power.multinomial(case, k=5, kappa=1)

[Package chisqalt version 0.0.0.900 Index]